powernode/dim/orchestrator/requirements.txt
# DIM Orchestrator Requirements

# Core dependencies
pydantic>=2.5.0
httpx>=0.25.0
pyyaml>=6.0.1

# IPFS integration (uses powernode.ipfs)
requests>=2.28.0

# Async support
asyncio-compat>=0.1.0

# Logging
# (uses standard library logging)

